+++ src/gpu_lib/__init__.py
@@ -1,2 +1,4 @@
 """Exposed gpu_lib classes and methods."""
 from gpu_lib.entities import Gpu
+from gpu_lib.errors import NotValidGpuError
+from gpu_lib.errors import DriverNotFound