+++ src/gpu_lib/errors/__init__.py
@@ -1,3 +1,3 @@
 """Exposed errors classes and methods."""
 from gpu_lib.errors.not_compatible_gpu_error import NotValidGpuError
-from gpu_lib.errors.not_valid_driver_error import NotValidDriverError
+from gpu_lib.errors.driver_not_found_error import DriverNotFound