+++ src/gpu_lib/errors/driver_not_found_error.py
@@ -1,9 +1,9 @@
 """This error represents if the system can't detect a valid GPU driver."""
 
 
-class NotValidDriverError(Exception):
+class DriverNotFound(Exception):
     """This error represents if the system can't detect a valid GPU driver."""
 
     def __init__(self) -> None:
         """Initialize the NotValidDriverError."""
-        super().__init__("The system can't detecto a valid driver.")
+        super().__init__("The system can't detect a valid GPU driver.")