+++ src/gpu_lib/managers/getter/__init__.py
@@ -1,3 +1,4 @@
 """Exposed nvidia classes and methods."""
 from gpu_lib.managers.getter.temperature_getter import TemperatureGetter
+from gpu_lib.managers.getter.use_getter import UseGetter
 from gpu_lib.managers.getter.nvidia import NvidiaTemperatureGetter