+++ src/gpu_lib/managers/getter/nvidia/nvidia_use_getter.py
@@ -1,18 +1,18 @@
-"""Obtain the temperature of GPU."""
+"""Obtain the use of GPU."""
 from shell_executor_lib import CommandError
 
-from gpu_lib.errors import DriverNotFound
-from gpu_lib.managers.getter import TemperatureGetter
+from gpu_lib import DriverNotFound
+from gpu_lib.managers.getter import UseGetter
 
 
-class NvidiaTemperatureGetter(TemperatureGetter):
-    """Obtain the temperature of GPU."""
+class NvidiaUseGetter(UseGetter):
+    """Obtain the use of GPU."""
 
-    async def get_temperature(self) -> float:
-        """Obtain the temperature of GPU in Celsius.
+    async def get_use(self) -> float:
+        """Obtain the use of GPU in percentage.
 
         Returns:
-            The temperature of GPU in Celsius.
+            The use of GPU.
 
         Raises:
             DriverNotFound: If the driver is not installed in the system.
@@ -20,7 +20,7 @@ class NvidiaTemperatureGetter(TemperatureGetter):
         """
         try:
             return float((await self._command_manager.execute_command(
-                "/bin/nvidia-smi --query-gpu=temperature.gpu --format=csv,noheader"))[0])
+                "/bin/nvidia-smi --query-gpu=temperature.gpu --format=csv,noheader"))[0].split(" ")[0])
         except CommandError as command_error:
             if command_error.status_code == 127:
                 raise DriverNotFound()