+++ tests/mock_gpu_lib/__init__.py
@@ -1,4 +1,4 @@
 """Exposed mock_gpu_lib classes and methods."""
-from tests.mock_gpu_lib.mock_nvidia import mock_temperature
+from tests.mock_gpu_lib.mock_nvidia import mock_temperature, mock_use_gpu
 
 mock_command_executor_method = "shell_executor_lib.CommandManager.execute_command"