+++ tests/mock_gpu_lib/mock_nvidia.py
@@ -1,3 +1,5 @@
 """Mocks of Nvidia GPU."""
 
 mock_temperature = ["50"]
+
+mock_use_gpu = ["20 %"]