+++ CHANGELOG.md
@@ -6,5 +6,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 ## Added
 - @javierfh03 - Develop the nvidia temperature getter with mocks and tests.
+- @javierfh03 - Develop the nvidia use getter with mocks and tests.
 
 [Unreleased]: https://github.com/Lagatrix/gpu_lib.git